src/old/requirements.txt
greenlet==3.2.2
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
SQLAlchemy==2.0.41
typing_extensions==4.13.2
